refactor(no-always-true-false-conditions): Enhance test suite for always true/false conditions
- Consolidate and improve test cases for the no-always-true-false-conditions rule
- Remove redundant test file for fix-related tests
- Merge mapped and rest invalid test cases into a single comprehensive test suite
- Improve test coverage and handling of complex condition scenarios
